Fixed precision bug
When the fog removed image was written to the backbuffer, it would cause the image to lose precision, and cause issues when fog was reintroduced. To combat this, the shader is now set up to store these truncated values and add them to the image when fog is being reintroduced. This consequently results the fog being more true to the original fog that was removed, and also results in better blending with any techniques added between removal and reintroduction.