Create ReVeil.fx
This shader is the newest version of fogremoval, and uses a new method that's referred to as guided Wiener filtering to refine the dark channel. The reason for the name change was to make it more accurate to its intended purpose, and so that I could leave the old fogremoval in the repository for those that still wish to use it.