Major update
Changed the filtering method so that now instead of just using the variance and mean to guide the filter, skewness and kurtosis can be used too. Almost artifacting issues in this build are gone now, and the second pass is no longer necessary. Any artifacting that remains is due to the image banding together or a small kernel size, and typically has an unnoticeable effect on the quality of the reintroduced image.